Experiment with device control
Each pipeline stage has an associated resource ID, and checks for device
control commands at the start of the stage. The first stage is given two
resource IDs, one representing the pipeline as a whole, the other for
the first stage.

Host app updated to talk to all four resource IDs and is working. Not
sure yet how we'll actually set this up, but this seems reasonable.