Experiment with multiple TinyUSB threads
Added a second call to usb_manager_start() to start two TinyUSB threads.
This is currently commented out as it may not be safe or necessary.

This allows the USB audio endpoint to keep moving while endpoint 0 is
blocked waiting for device control which can take up to a pipeline frame
time (15ms) to complete.